Refactored a bit the security groups
As there's normally a limit of 5 security groups that can be
attached to an instance, and we're deploying both Teleport services
`proxy` and `auth` together, I've created a combined security group
`bastion` with all the required rules to run all Teleport services
together. There's still a separate security group for the Teleport
nodes.